Added profile function, Added userid & videoid into Models, Created uploading videos and avatar

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async(req, res) => {
     try {
@@ -10,43 +11,68 @@ export const home = async(req, res) => {
 }
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video) {
         return res.status(404).render("404", {pageTitle:"Video Not Found"});
     } 
     return res.render("videos/watch", {pageTitle: video.title, video});
 };
 export const getEdit = async (req, res) => {
-    const { id } = req.params;
+    const {
+        params: { id },
+        session: { user: {_id}},
+     } = req;
     const video = await Video.findById(id);
     if(!video) {
         return res.status(404).render("404", {pageTitle:"Video Not Found"});
     } 
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     return res.render("videos/edit", {pageTitle: `Editing ${video.title}`, video});
 };
 export const postEdit = async (req, res) => {
-    const { id } = req.params;
-    const {title, description, hashtags} = req.body;
-    if(!Video.exists({_id:id})) {
+    const {
+        params: { id },
+        body: {title, description, hashtags},
+        session: {
+            user: { _id },
+        },
+    } = req;
+    const video = await Video.findById(id);
+    if(!video) {
         return res.status(404).render("404", {pageTitle:"Video Not Found"});
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title, description, hashtags: Video.formatHashtags(hashtags),
     })
-    return res.redirect(`videos/${id}`);
+    
+    return res.redirect(`/videos/${id}`);
 };
 
 export const getUpload = (req, res) => {
     return res.render("videos/upload", {pageTitle: "Upload Video"});
 };
 export const postUpload = async (req, res) => {
-    const { title, description, hashtags } = req.body;
+    const {
+        session: { user:{_id} },
+        body: {title, description, hashtags },
+        file,
+    } = req;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl: file ? file.path : "",
             hashtags: Video.formatHashtags(hashtags),
+            owner: _id,
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch(error) {
         console.log(error);
@@ -55,7 +81,15 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-    const { id } = req.params;
+    const {
+        params: { id },
+        session: { user: {_id}},
+    } = req;
+    const video = await Video.exists({_id: id});
+    console.log(video);
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
